Split config and restructure file
