+++ helpers/contentIndexBuilder/index.js
@@ -1,0 +1,19 @@
+const yaml = require('js-yaml')
+const fs = require('fs')
+
+const config = require('./config')
+
+try {
+  const contentIndex = config.reduce((prevContent, content) => [
+    ...prevContent,
+    ...yaml.safeLoad(fs.readFileSync(content.path, 'utf8')).map(record => ({
+      q: content.getQ(record),
+      type: content.type,
+      url: content.getUrl(record),
+    }))
+  ], [])
+
+  fs.writeFileSync('./src/contents/index.yaml', yaml.safeDump(contentIndex), 'utf8')
+} catch (e) {
+  console.log(e)
+}
